+++ application/Forms/Rss/FeedForm.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: Rhilip
+ * Date: 6/6/2020
+ * Time: 8:09 PM
+ */
+
+declare(strict_types=1);
+
+namespace App\Forms\Rss;
+
+use App\Forms\Torrents\SearchForm;
+
+class FeedForm extends SearchForm
+{
+}