+++ templates/manage/categories.php
@@ -5,8 +5,10 @@
  * Date: 7/15/2019
  * Time: 10:40 PM
  * @var League\Plates\Template\Template $this
- * @var array $categories
+ * @var \App\Forms\Manage\Categories\IndexForm $form
  */
+
+$categories = $form->getCategories();
 ?>
 
 <?= $this->layout('layout/base') ?>
@@ -37,16 +39,15 @@
             <?php if (count($categories) > 0): ?>
                 <?php foreach ($categories as $category): ?>
                     <tr data-id="<?= $category['id'] ?>"
-                        data-parent_id="<?= $category['parent_id'] ?>"
                         data-enabled="<?= $category['enabled'] ?>"
                         data-name="<?= $category['name'] ?>"
-                        data-full_name="<?= $category['full_path'] ?>"
                         data-image="<?= $category['image'] ?>"
                         data-class_name="<?= $category['class_name'] ?>"
+                        data-sort_index="<?= $category['sort_index'] ?>"
                     >
                         <td><?= $category['enabled'] ? '<i class="far fa-fw fa-check-square"></i>':'<i class="far fa-fw fa-square"></i>' ?></td>
                         <td><?= $category['id'] ?></td>
-                        <td><?= str_repeat('&nbsp;', 4 * ($category['level'] + 1)) ?><b><?= $category['name'] ?></b></td>
+                        <td><b><?= $category['name'] ?></b></td>
                         <td><?= $category['image'] ?></td>
                         <td><?= $category['class_name'] ?></td>
                         <td><a href="javascript:" class="cat-edit" data-id="<?= $category['id'] ?>">Edit</a> |
@@ -66,54 +67,76 @@
 <?php $this->end(); ?>
 
 <?php $this->start('body');?>
-<form method="post" id="cat_remove_form" class="hidden">
-    <label><input type="text" name="action" value="cat_delete"></label>
-    <label><input type="number" name="cat_id" value=""></label>
-</form>
-
-<div class="modal fade" id="cat_modal">
+<div class="modal fade" id="cat_delete_modal">
     <div class="modal-dialog">
         <div class="modal-content">
             <div class="modal-header">
                 <button type="button" class="close" data-dismiss="modal"><span aria-hidden="true">×</span><span class="sr-only">Close</span></button>
                 <h4 class="modal-title">Add/Edit Category</h4>
             </div>
-            <form method="post" class="form-horizontal" id="cat_edit_form" data-toggle="validator" role="form">
+            <form action="/manage/categories/delete" method="post" class="form-horizontal" id="cat_remove_form" data-toggle="validator" role="form">
                 <div class="modal-body">
-                    <label class="hidden"><input type="text" name="action" value="cat_edit"></label>
-                    <label class="hidden"><input type="number" id="cat_id" name="cat_id" value="0"></label>
+                <label class="hidden"><input type="number" id="id" name="id" value="0"></label>
                     <div class="form-group">
-                        <label for="cat_name" class="required">Name</label>
-                        <input type="text" class="form-control" id="cat_name" name="cat_name" required>
+                        <label for="move_to" class="required">Move To</label>
+                        <select class="form-control" id="move_to" name="move_to">
+                            <option value="">[Choose One Category]</option>
+                            <?php foreach (container()->get('site')->ruleCategory() as $id => $category): ?>
+                                <option value="<?= $id ?>"><?= $category['name'] ?></option>
+                            <?php endforeach; ?>
+                        </select>
+                        <div class="help-block">Since you delete this Category, You MUST set another category for
+                            torrent to update.
+                        </div>
+                    </div>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-default" id="cat_modal_close" data-dismiss="modal">Close</button>
+                    <button type="submit" class="btn btn-primary" id="cat_modal_save">Save</button>
+                </div>
+            </form>
+        </div>
+    </div>
+</div>
+
+
+
+<div class="modal fade" id="cat_edit_modal">
+    <div class="modal-dialog">
+        <div class="modal-content">
+            <div class="modal-header">
+                <button type="button" class="close" data-dismiss="modal"><span aria-hidden="true">×</span><span class="sr-only">Close</span></button>
+                <h4 class="modal-title">Add/Edit Category</h4>
+            </div>
+            <form action="/manage/categories/edit" method="post" class="form-horizontal" id="cat_edit_form" data-toggle="validator" role="form">
+                <div class="modal-body">
+                    <label class="hidden"><input type="number" id="id" name="id" value="0"></label>
+                    <div class="form-group">
+                        <label for="name" class="required">Name</label>
+                        <input type="text" class="form-control" id="name" name="name" required>
                         <div class="help-block">Don't use long name. Recommend less than 10 letters.</div>
                     </div>
                     <div class="form-group">
                         <div class="switch text-left">
-                            <input type="checkbox" id="cat_enabled" name="cat_enabled" value="1" checked>
-                            <label for="cat_enabled">Enabled</label>
+                            <input type="checkbox" id="enabled" name="enabled" value="1" checked>
+                            <label for="enabled">Enabled</label>
                         </div>
                     </div>
                     <div class="form-group">
-                        <label for="cat_parent_id" class="required">Parent Category</label>
-                        <select id="cat_parent_id" name="cat_parent_id" class="form-control">
-                            <?php foreach ($categories as $category) : ?>
-                                <?php $level = substr_count($category['full_path'], ' - '); ?>
-                                <option value="<?= $category['id'] ?>"><?= str_repeat('&nbsp;', 4 * ($category['level'] + 1)) .  $category['name'] ?></option>
-                            <?php endforeach; ?>
-                        </select>
-                        <div class="help-block">Recommended up to two levels of category.</div>
-                    </div>
-                    <div class="form-group">
-                        <label for="cat_image">Image</label>
-                        <input type="text" class="form-control" id="cat_image" name="cat_image"
+                        <label for="image">Image</label>
+                        <input type="text" class="form-control" id="image" name="image"
                                pattern="^[a-z0-9_./]*$">
-                        <div class="help-block">The name of image file.Leave it blank to show Text Only. <code>Allowed Characters: [a-z] (in lower case), [0-9], [_./].</code></div>
+                        <div class="help-block">The name of image file.Leave it blank to show Text Only. Allowed Characters: <code>[a-z] (in lower case), [0-9], [_./].</code></div>
                     </div>
                     <div class="form-group">
-                        <label for="cat_class_name">Class Name</label>
-                        <input type="text" class="form-control" id="cat_class_name" name="cat_class_name"
+                        <label for="class_name">Class Name</label>
+                        <input type="text" class="form-control" id="class_name" name="class_name"
                                pattern="^[a-z][a-z0-9_\-]*?$">
-                        <div class="help-block">The value of 'class' attribute of the image. Leave it blank if none. <code>Allowed Characters: [a-z] (in lower case), [0-9], [-_], and the first letter must be in [a-z].</code></div>
+                        <div class="help-block">The value of 'class' attribute of the image. Leave it blank if none. Allowed Characters: <code>[a-z] (in lower case), [0-9], [-_], and the first letter must be in [a-z].</code></div>
+                    </div>
+                    <div class="form-group">
+                        <label for="sort_index">Sort Index</label>
+                        <input type="number" class="form-control" id="sort_index" name="sort_index" value="0">
                     </div>
                 </div>
                 <div class="modal-footer">